+++ examples/Go/package.json
