+++ examples/Nodejs/hello.js
@@ -1,0 +1,6 @@
+module.exports = async function (context) {
+    return {
+        status: 200,
+        body: "Hello, world!\n"
+    };
+}
