+++ examples/Nodejs/serverless.yml
@@ -1,0 +1,13 @@
+service: hello
+
+provider:
+  name: fission
+  runtime: nodejs
+
+plugins:
+  - serverless-fission
+
+functions:
+  hello:
+    description: 'Hello function'
+    handler: handler.hello
