+++ examples/Python/package.json
@@ -1,0 +1,16 @@
+
+{
+  "name": "hello",
+  "version": "1.0.0",
+  "description": "Example function for serverless kubeless",
+  "dependencies": {
+    "node-run-cmd": "^1.0.1",
+    "serverless-fission": "^0.0.1"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "Apache-2.0"
+}
