--- hello.js
@@ -1,7 +1,0 @@
-
-module.exports = async function(context) {
-    return {
-        status: 200,
-        body: "Hello, world!\n"
-    };
-}
