+++ package.json
@@ -28,7 +28,7 @@
     "eslint-config-airbnb-base": "^5.0.2",
     "eslint-config-standard": "^11.0.0",
     "eslint-plugin-async": "^0.1.1",
-    "eslint-plugin-import": "^1.16.0",
+    "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-jsx-a11y": "^2.1.0",
     "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.8.0",
