+++ print/fissionPrint.js
@@ -13,10 +13,10 @@
 'use strict';
 const Bbpromise = require('bluebird');
 var nrc = require('node-run-cmd');
-var fs = require('fs');
-path = require('path');
-const Client = require('kubernetes-client').Client;
-const config = require('kubernetes-client').config;
+//var fs = require('fs');
+//path = require('path');
+//const Client = require('kubernetes-client').Client;
+//const config = require('kubernetes-client').config;
 class fissionPrint {
 	constructor(serverless,options) {
 	this.serverless = serverless;
@@ -29,7 +29,7 @@ class fissionPrint {
 		
 	print() {
 	var errorCallback = function(data) {
-		  console.log(data);
+		console.log(data);
 	};
 	var dataCallback=function(data) {
 		console.log(data);
